User login stats, test stats component rename, additional user stats methods on controller. Pre login stats refactor
